requirements.txt
PyYAML==3.13
tensorflow==1.12.0
Pillow==5.3.0